feat: added character count for prompt and response to enable cost monitoring. refactor: moved all Vertex AI components for Client and API to its own module.
