+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "google-generative-ai-rs"
-version = "0.2.2"
+version = "0.2.3"
 edition = "2021"
 authors = ["Mick Clarke <avastmick@outlook.com>"]
 license = "MIT"
@@ -13,6 +13,7 @@ keywords = ["google", "generative", "ai", "gemini", "client"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bytecount = "0.6.7"
 env_logger = { version = "0.11" }
 futures = { version = "0.3" }
 gcp_auth = { version = "0.10" }
