+++ src/v1/mod.rs
@@ -2,3 +2,4 @@ pub mod api;
 pub mod errors;
 pub mod gemini;
 pub mod tools;
+pub mod vertexai;
